pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.97.1",
    "pinecone>=7.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
